+++ src/@lib/hooks/useCallback.ts
@@ -1,8 +1,9 @@
+/* eslint-disable @typescript-eslint/no-unsafe-function-type */
 /* eslint-disable react-hooks/exhaustive-deps */
 import { DependencyList } from "react";
 import { useMemo } from "./useMemo";
 
-export function useCallback<T extends () => T>(
+export function useCallback<T extends Function>(
   factory: T,
   _deps: DependencyList,
 ) {
