+++ src/components/index.ts
@@ -1,0 +1,6 @@
+import Header from "./Header";
+import NotificationSystem from "./NotificationSystem";
+import ComplexForm from "./ComplexForm";
+import ItemList from "./ItemList";
+
+export { Header, NotificationSystem, ComplexForm, ItemList };
