+++ src/types/context.ts
@@ -1,0 +1,18 @@
+import { Notification, User } from "./common";
+
+export interface ThemeContextType {
+  theme: string;
+  toggleTheme: () => void;
+}
+
+export interface UserContextType {
+  user: User | null;
+  login: (email: string, password: string) => void;
+  logout: () => void;
+}
+
+export interface NotificationContextType {
+  notifications: Notification[];
+  addNotification: (message: string, type: Notification["type"]) => void;
+  removeNotification: (id: number) => void;
+}
